requirements.txt
requests
PyYAML
watchdog
tabulate